Refine WH851 Bluetooth GATT backend structure
